+++ src/components/client/core/index.ts
@@ -2,6 +2,7 @@ export { default as Accordion } from "./Accordion/Accordion";
 export { default as Banner } from "./Banner/Banner";
 export { default as Breadcrumb } from "./Breadcrumb/Breadcrumb";
 export { default as Button } from "./Button/Button";
+export { default as Carousel } from "./Carousel/Carousel";
 export { default as Checkbox } from "./Checkbox/Checkbox";
 export { default as Drawer } from "./Drawer/Drawer";
 export { default as Icon } from "./Icon/Icon";
@@ -24,6 +25,7 @@ export type {
 export type { Props as BannerProps } from "./Banner/Banner";
 export type { Props as BreadcrumbProps } from "./Breadcrumb/Breadcrumb";
 export type { Props as ButtonProps } from "./Button/Button";
+export type { Props as CarouselProps } from "./Carousel/Carousel";
 export type { Props as CheckboxProps } from "./Checkbox/Checkbox";
 export type { Props as DrawerProps } from "./Drawer/Drawer";
 export type { Props as IconProps } from "./Icon/Icon";