+++ src/lib/hooks/useDisclosure/useDisclosure.stories.tsx
@@ -3,11 +3,12 @@ import { Text } from "components/universal";
 import { useDisclosure } from "lib/hooks";
 
 import type { Meta, StoryObj } from "@storybook/react";
+import type { ComponentType } from "react";
 
-type Story = StoryObj<typeof Modal>;
+type Story = StoryObj<typeof useDisclosure>;
 
 const DisclosureExample = () => {
-  const { isOpen, onOpen, onClose } = useDisclosure({});
+  const { isOpen, onOpen, onClose } = useDisclosure();
 
   return (
     <Modal
@@ -28,11 +29,12 @@ export const Default: Story = {
 };
 
 // TODO remove explicit type annotation, required due to `pnpm` bug (and therefore Yarn with `pnpm` linker); https://github.com/microsoft/TypeScript/issues/47663
-const meta: Meta<typeof Modal> = {
+const meta: Meta<typeof useDisclosure> = {
   title: "Hooks/useDisclosure",
-  component: Modal,
   tags: ["autodocs"],
+  // NB: type coercion here to allow `useDisclosure` Storybook metadata to render (e.g. JSDoc, hook parameters)
+  component: useDisclosure as unknown as ComponentType,
   decorators: [(Story) => <Story />],
-} satisfies Meta<typeof Modal>;
+} satisfies Meta<typeof useDisclosure>;
 
 export default meta;